Implement `figureVideoElement` component
