+++ packages/11ty/_includes/components/index.js
@@ -28,6 +28,7 @@ module.exports = {
   figureTable: require('./figure/table'),
   figureTableElement: require('./figure/table/element'),
   figureVideo: require('./figure/video'),
+  figureVideoElement: require('./figure/video/element'),
   head: require('./head.js'),
   icon: require('./icon.js'),
   icons: require('./icons.js'),
